+++ packages/core/scripts/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./docgen";
 export * from "./packageBuilder";
+export * from "./releaseBuilder";
+export * from "./releaseLauncher";